Rearrange validate module to single file based
